rename select input and listbox input
