+++ module/src/index.ts
@@ -18,8 +18,8 @@ export * from './components/iconButton';
 export * from './components/input';
 export * from './components/inputWrapper';
 export * from './components/nativeDateInput';
-export * from './components/nativeSelectInput';
 export * from './components/numberInput';
+export * from './components/select';
 export * from './components/selectInput';
 export * from './components/spinner';
 export * from './components/status';
